Refactor midlewares
