+++ safe_transaction_service/middlewares.py
@@ -1,6 +1,44 @@
 import logging
 
-logger = logging.getLogger(__name__)
+from loggers.custom_logger import (
+    HttpRequest,
+    HttpResponseLog,
+    get_milliseconds_now,
+    http_request_log,
+)
+
+
+class LoggingMiddleware:
+    """
+    Http Middleware to generate request and response logs.
+    """
+
+    def __init__(self, get_response):
+        self.get_response = get_response
+        self.logger = logging.getLogger("LoggingMiddleware")
+
+    def __call__(self, request: HttpRequest):
+        start_time = get_milliseconds_now()
+        response = self.get_response(request)
+        if request.resolver_match:
+            end_time = get_milliseconds_now()
+            delta = end_time - start_time
+            http_request = http_request_log(request, start_time)
+            content: str | None = None
+            if 400 <= response.status_code < 500 and hasattr(response, "data"):
+                content = str(response.data)
+
+            http_response = HttpResponseLog(
+                response.status_code, end_time, delta, content
+            )
+            self.logger.info(
+                "Http request",
+                extra={
+                    "http_response": http_response,
+                    "http_request": http_request,
+                },
+            )
+        return response
 
 
 class ProxyPrefixMiddleware:
@@ -13,10 +51,11 @@ class ProxyPrefixMiddleware:
 
     def __init__(self, get_response):
         self.get_response = get_response
+        self.logger = logging.getLogger("ProxyPrefixMiddleware")
 
     def __call__(self, request):
         prefix = request.META.get("HTTP_X_FORWARDED_PREFIX", "")
-        logger.debug(f"HTTP_X_FORWARDED_PREFIX:{prefix}")
+        self.logger.debug(f"HTTP_X_FORWARDED_PREFIX:{prefix}")
         if prefix:
             original_get_full_path = request.get_full_path
